Refactored request processors to specific packages
