+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/ClassExclusionFilter.scala
@@ -1,7 +1,7 @@
 package org.senkbeil.debugger.jdi.requests.filters
 
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.ClassExclusionProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.ClassExclusionProcessor
 
 /**
  * Represents a filter used to limit requests to any class not specified
