+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/ClassInclusionFilter.scala
@@ -1,7 +1,7 @@
 package org.senkbeil.debugger.jdi.requests.filters
 
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.ClassInclusionProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.ClassInclusionProcessor
 
 /**
  * Represents a filter used to limit requests to only classes specified by
