+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/ClassReferenceFilter.scala
@@ -2,7 +2,7 @@ package org.senkbeil.debugger.jdi.requests.filters
 
 import com.sun.jdi.ReferenceType
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.ClassReferenceProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.ClassReferenceProcessor
 
 /**
  * Represents a filter used to limit requests to a specific class reference.
