+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/CountFilter.scala
@@ -1,7 +1,7 @@
 package org.senkbeil.debugger.jdi.requests.filters
 
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.CountProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.CountProcessor
 
 /**
  * Represents a filter used to restrict events until the specific event has
