+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/InstanceFilter.scala
@@ -2,7 +2,7 @@ package org.senkbeil.debugger.jdi.requests.filters
 
 import com.sun.jdi.ObjectReference
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.InstanceProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.InstanceProcessor
 
 /**
  * Represents a filter used to limit requests to a specific instance of a class.
