+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/SourceNameFilter.scala
@@ -1,7 +1,7 @@
 package org.senkbeil.debugger.jdi.requests.filters
 
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.SourceNameProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.SourceNameProcessor
 
 /**
  * Represents a filter used to limit requests to reference types with whom a
