+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/ThreadFilter.scala
@@ -2,7 +2,7 @@ package org.senkbeil.debugger.jdi.requests.filters
 
 import com.sun.jdi.ThreadReference
 import org.senkbeil.debugger.jdi.requests.JDIRequestProcessor
-import org.senkbeil.debugger.jdi.requests.processors.ThreadProcessor
+import org.senkbeil.debugger.jdi.requests.filters.processors.ThreadProcessor
 
 /**
  * Represents a filter used to limit requests to a specific thread.
