+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/processors/ClassReferenceProcessor.scala
@@ -1,9 +1,9 @@
-package org.senkbeil.debugger.jdi.requests.processors
+package org.senkbeil.debugger.jdi.requests.filters.processors
 
 import com.sun.jdi.ReferenceType
 import com.sun.jdi.request._
-import org.senkbeil.debugger.jdi.requests.{JDIRequestProcessor, JDIRequestArgument}
 import org.senkbeil.debugger.jdi.requests.filters.ClassReferenceFilter
+import org.senkbeil.debugger.jdi.requests.{JDIRequestArgument, JDIRequestProcessor}
 
 /**
  * Represents a processor for the class reference filter.
