+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/processors/CountProcessor.scala
@@ -1,8 +1,8 @@
-package org.senkbeil.debugger.jdi.requests.processors
+package org.senkbeil.debugger.jdi.requests.filters.processors
 
 import com.sun.jdi.request.EventRequest
-import org.senkbeil.debugger.jdi.requests.{JDIRequestProcessor, JDIRequestArgument}
 import org.senkbeil.debugger.jdi.requests.filters.CountFilter
+import org.senkbeil.debugger.jdi.requests.{JDIRequestArgument, JDIRequestProcessor}
 
 /**
  * Represents a processor for the count filter.
