+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/filters/processors/InstanceProcessor.scala
@@ -1,9 +1,9 @@
-package org.senkbeil.debugger.jdi.requests.processors
+package org.senkbeil.debugger.jdi.requests.filters.processors
 
 import com.sun.jdi.ObjectReference
 import com.sun.jdi.request._
-import org.senkbeil.debugger.jdi.requests.{JDIRequestProcessor, JDIRequestArgument}
 import org.senkbeil.debugger.jdi.requests.filters.InstanceFilter
+import org.senkbeil.debugger.jdi.requests.{JDIRequestArgument, JDIRequestProcessor}
 
 /**
  * Represents a processor for the instance filter.
