+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/jdi/requests/properties/processors/EnabledProcessor.scala
@@ -1,8 +1,8 @@
-package org.senkbeil.debugger.jdi.requests.processors
+package org.senkbeil.debugger.jdi.requests.properties.processors
 
 import com.sun.jdi.request._
-import org.senkbeil.debugger.jdi.requests.{JDIRequestProcessor, JDIRequestArgument}
-import org.senkbeil.debugger.jdi.requests.properties.{EnabledProperty, CustomProperty}
+import org.senkbeil.debugger.jdi.requests.properties.EnabledProperty
+import org.senkbeil.debugger.jdi.requests.{JDIRequestArgument, JDIRequestProcessor}
 
 /**
  * Represents a processor for the enabled status of a request.
