+++ scala-debugger-api/src/test/scala/org/senkbeil/debugger/jdi/requests/filters/processors/ClassReferenceProcessorSpec.scala
@@ -1,4 +1,4 @@
-package org.senkbeil.debugger.jdi.requests.processors
+package org.senkbeil.debugger.jdi.requests.filters.processors
 
 import com.sun.jdi.ReferenceType
 import com.sun.jdi.request._
